+++ featurebyte/api/view.py
@@ -22,6 +22,7 @@ from featurebyte.api.join_utils import (
 )
 from featurebyte.core.frame import Frame
 from featurebyte.core.generic import ProtectedColumnsQueryObject
+from featurebyte.core.mixin import SampleMixin
 from featurebyte.core.series import Series
 from featurebyte.core.util import append_to_lineage
 from featurebyte.exception import NoJoinKeyFoundError, RepeatedColumnNamesError
@@ -38,7 +39,7 @@ else:
 ViewT = TypeVar("ViewT", bound="View")
 
 
-class ViewColumn(Series):
+class ViewColumn(Series, SampleMixin):
     """
     ViewColumn class that is the base class of columns returned from any View (e.g. EventView)
     """
