+++ featurebyte/schema/scd_data.py
@@ -18,7 +18,7 @@ class SCDDataCreate(DataCreate):
     """
 
     natural_key_column: StrictStr
-    surrogate_key_column: StrictStr
+    surrogate_key_column: Optional[StrictStr]
     effective_timestamp_column: StrictStr
     end_timestamp_column: Optional[StrictStr]
     current_flag: Optional[StrictStr]
