+++ tests/unit/routes/test_feature_store.py
@@ -2,11 +2,15 @@
 Test for FeatureStore route
 """
 from http import HTTPStatus
+from unittest.mock import Mock
 
+import pandas as pd
 import pytest
 from bson.objectid import ObjectId
+from pandas.testing import assert_frame_equal
 
 from featurebyte.exception import CredentialsError
+from featurebyte.schema.feature_store import FeatureStoreSample
 from tests.unit.routes.base import BaseApiTestSuite
 
 
@@ -250,3 +254,118 @@ class TestFeatureStoreApi(BaseApiTestSuite):
             {"name": "b", "dtype": "INT"},
             {"name": "c", "dtype": "BOOL"},
         ]
+
+    @pytest.fixture(name="data_sample_payload")
+    def data_sample_payload_fixture(
+        self, test_api_client_persistent, create_success_response, snowflake_feature_store
+    ):
+        """Payload for data sample"""
+        _ = create_success_response
+        test_api_client, _ = test_api_client_persistent
+        payload = self.load_payload("tests/fixtures/request_payloads/event_data.json")
+        response = test_api_client.post("/event_data", json=payload)
+        assert response.status_code == HTTPStatus.CREATED, response.json()
+
+        data_response_dict = response.json()
+        return FeatureStoreSample(
+            feature_store_name=snowflake_feature_store.name,
+            graph={
+                "edges": [],
+                "nodes": [
+                    {
+                        "name": "input_1",
+                        "output_type": "frame",
+                        "parameters": {
+                            "type": data_response_dict["type"],
+                            "id": "6332fdb21e8f0eeccc414512",
+                            "columns": [
+                                "col_int",
+                                "col_float",
+                                "col_char",
+                                "col_text",
+                                "col_binary",
+                                "col_boolean",
+                                "event_timestamp",
+                                "created_at",
+                                "cust_id",
+                            ],
+                            "table_details": {
+                                "database_name": "sf_database",
+                                "schema_name": "sf_schema",
+                                "table_name": "sf_table",
+                            },
+                            "feature_store_details": snowflake_feature_store.json_dict(),
+                        },
+                        "type": "input",
+                    },
+                ],
+            },
+            node_name="input_1",
+            from_timestamp="2012-11-24T11:00:00",
+            to_timestamp="2019-11-24T11:00:00",
+            timestamp_column="event_timestamp",
+        ).json_dict()
+
+    def test_sample_200(self, test_api_client_persistent, data_sample_payload, mock_get_session):
+        """Test data preview (success)"""
+        test_api_client, _ = test_api_client_persistent
+
+        expected_df = pd.DataFrame({"a": [0, 1, 2]})
+        mock_session = mock_get_session.return_value
+        mock_session.execute_query.return_value = expected_df
+        mock_session.generate_session_unique_id = Mock(return_value="1")
+        response = test_api_client.post("/feature_store/sample", json=data_sample_payload)
+        assert response.status_code == HTTPStatus.OK
+        assert_frame_equal(pd.read_json(response.json(), orient="table"), expected_df)
+        assert mock_session.execute_query.call_args[0][0].endswith(
+            "WHERE\n  event_timestamp >= CAST('2012-11-24T11:00:00' AS TIMESTAMP)\n  "
+            "AND event_timestamp < CAST('2019-11-24T11:00:00' AS TIMESTAMP)\n"
+            "ORDER BY\n  RANDOM(1234)\nLIMIT 10"
+        )
+
+    def test_sample_422__no_timestamp_column(self, test_api_client_persistent, data_sample_payload):
+        """Test data preview no timestamp column"""
+        test_api_client, _ = test_api_client_persistent
+        response = test_api_client.post(
+            "/feature_store/sample",
+            json={
+                **data_sample_payload,
+                "from_timestamp": "2012-11-24T11:00:00",
+                "to_timestamp": None,
+                "timestamp_column": None,
+            },
+        )
+        assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
+        assert response.json() == {
+            "detail": [
+                {
+                    "loc": ["body", "__root__"],
+                    "msg": "timestamp_column must be specified.",
+                    "type": "value_error",
+                }
+            ]
+        }
+
+    def test_sample_422__invalid_timestamp_range(
+        self, test_api_client_persistent, data_sample_payload
+    ):
+        """Test data preview no timestamp column"""
+        test_api_client, _ = test_api_client_persistent
+        response = test_api_client.post(
+            "/feature_store/sample",
+            json={
+                **data_sample_payload,
+                "from_timestamp": "2012-11-24T11:00:00",
+                "to_timestamp": "2012-11-20T11:00:00",
+            },
+        )
+        assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
+        assert response.json() == {
+            "detail": [
+                {
+                    "loc": ["body", "__root__"],
+                    "msg": "from_timestamp must be smaller than to_timestamp.",
+                    "type": "assertion_error",
+                }
+            ]
+        }
